Move duplicate code in pet data fetch requests to PetDataFetchActionRequestModel
