+++ app/src/main/java/com/example/cupetfrontend/use_cases/request_models/pet/GetPotentialMatchesRequestModel.java
@@ -1,19 +1,7 @@
 package com.example.cupetfrontend.use_cases.request_models.pet;
 
-public class GetPotentialMatchesRequestModel {
-    private final String token;
-    private final String petId;
-
+public class GetPotentialMatchesRequestModel extends PetDataFetchActionRequestModel{
     public GetPotentialMatchesRequestModel(String token, String petId) {
-        this.token = token;
-        this.petId = petId;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public String getPetId() {
-        return petId;
+        super(token, petId);
     }
 }
